+++ src/components/Shop/DetailPage/ContentWrapper/DetailImage/index.tsx
