+++ src/components/Shop/DetailPage/ContentWrapper/Notice/index.tsx
